Move `ctp connect` to `ctp connector install`
